--- a/OpenSim/Region/Framework/Interfaces/ITerrain.cs
+++ b/OpenSim/Region/Framework/Interfaces/ITerrain.cs
@@ -73,6 +73,6 @@ namespace OpenSim.Region.Framework.Interfaces
 
     public interface IMapImageGenerator
     {
-        byte[] WriteJpeg2000Image(string gradientmap);
+        byte[] WriteJpeg2000Image();
     }
 }