… OpenSim/Region/CoreModules/World/WorldMap/ShadedMapTileRenderer.cs
… OpenSim/Region/CoreModules/World/WorldMap/TexturedMapTileRenderer.cs
… OpenSim/Region/CoreModules/World/WorldMap/WorldMapModule.cs
… OpenSim/Region/Framework/Interfaces/IWorldMapModule.cs
… OpenSim/Region/Framework/Scenes/Scene.cs
… OpenSim/Region/Framework/Scenes/SceneBase.cs
… OpenSim/Region/Framework/Scenes/SceneGraph.cs
… OpenSim/Region/Framework/Scenes/SceneObjectPart.cs
… bin/Halcyon.sample.ini