Draws the objects correctly as far as it goes: the colors used are per-face instead of per-object, they are NOT inverted, and OBBs are scaled correctly, not 2X.

Before any optimization:

```
06:13:23 - [MAPTILE]: Generating Maptile Step 2: Done in 920 ms, Filtering/Prepping took 514, Array cloning took 1, Sorting took 15, Drawing took 390 (=1840ms)
06:14:24 - [MAPTILE]: Generating Maptile Step 2: Done in 839 ms, Filtering/Prepping took 433, Array cloning took 0, Sorting took 15, Drawing took 391 (=1678ms)
06:15:26 - [MAPTILE]: Generating Maptile Step 2: Done in 1315 ms, Filtering/Prepping took 468, Array cloning took 453, Sorting took 18, Drawing took 376 (=2,630ms)
```

After this patch (with the metrics uncommented of course)

```
05:56:38 - [MAPTILE]: Generating Maptile Step 2 (Objects): Processed 44743 entities, 44992 prims, 44727 used for map drawing, resulting in 44740 faces to draw.
05:56:38 - [MAPTILE]: Generating Maptile Step 2 (Objects): Timing: prepping took 6ms, filtering prims took 87ms, calculating vertices took 66ms, computing sorting height took 14ms, calculating OBB normal took 31ms, getting OBB face colors took 114ms, calculating OBBs took 17ms, adding OBBs to list took 10ms, sorting took 78ms, drawing took 94ms, total time: 745ms
05:56:50 - [MAPTILE]: Generating Maptile Step 2 (Objects): Processed 44743 entities, 44992 prims, 44727 used for map drawing, resulting in 44740 faces to draw.
05:56:50 - [MAPTILE]: Generating Maptile Step 2 (Objects): Timing: prepping took 6ms, filtering prims took 94ms, calculating vertices took 62ms, computing sorting height took 18ms, calculating OBB normal took 26ms, getting OBB face colors took 98ms, calculating OBBs took 24ms, adding OBBs to list took 5ms, sorting took 87ms, drawing took 94ms, total time: 775ms
05:57:02 - [MAPTILE]: Generating Maptile Step 2 (Objects): Processed 44743 entities, 44992 prims, 44727 used for map drawing, resulting in 44740 faces to draw.
05:57:02 - [MAPTILE]: Generating Maptile Step 2 (Objects): Timing: prepping took 6ms, filtering prims took 83ms, calculating vertices took 62ms, computing sorting height took 14ms, calculating OBB normal took 41ms, getting OBB face colors took 99ms, calculating OBBs took 22ms, adding OBBs to list took 11ms, sorting took 83ms, drawing took 94ms, total time: 761ms
```