@appurist 2017-02-10
Just reviewing some of these older PRs again.  The changes for these tile optimizations look good, except I'd feel more at ease about them if the timer wasn't restarted _before_ the processing started, but instead was restarted after it completed.  WorldMapModule.cs line 1173:
https://github.com/InWorldz/halcyon/pull/257/files#diff-a9e3bf7a3b06e734944bda023851354cR1173

I only have one slight concern, and that is that the minimum time should be the minimum _between_ runs, not the minimum between start times. i.e. excluding processing time.  If it's set to 60-second refresh, and it takes 20 seconds to process, there should still be 60 seconds of idle time between those, not 40 seconds (the difference).  Just in case something goes wrong somewhere, and it ends up taking a minute (or more) and it's set to refresh every minute (that would be a tight loop then). If it's the minimum time between, it guarantees at least that much idle time.