@kf6kjg 2017-02-10
The big question is whether or not the entire operation should be removed from the simulator: separation of concerns etc.  This section of code is only for generating map tiles, not even the minimap uses this.