@appurist 2017-02-11
That doesn't actually address my concern, which admittedly was a bit academic. Mostly that the minimum wait time is not really a minimum _wait_ time, and I know the usual processing time is measured in seconds.  It's just that by moving the `timer.Start` a couple of lines later, it avoids all of this, and provides for scalability beyond the current map sizes.  (Just ask yourself if this handles 100 times the data. No, but it would with a one-line move.  And even with 10 times the data, it would be using about 50% of a processor, plus or minus a lot since we don't really have a good handle on how long a run will take.)  I was just pointing out that these considerations were moot if the minimum wait time was the time between the _end_ of processing and the start of the next run.