@kf6kjg 2017-02-11
Ok, so maybe I should have typed "semi-address" :P

I agree that doing so would address the rendering phase delay scaling issue your bring up.