@kf6kjg 2017-05-14
While Anaximander 2 generates its own map tiles now, the Halcyon server still has the capacity to do so, so I'm leaving this open until such a time as its either merged or it's decided that the map tile generation subsystem is to be removed.