@appurist 2017-06-20
I'm going to merge this for now but if Anaximander 2 makes this obsolete, I won't shed any tears. ;)